Add files via upload
These label macros use outlines and shadows to create labels on images that stand out against the image underneath.

 The Fancy Scale Bar macro adds extensive formatting options to the original version by Wayne Rasband that was subsequently enhanced by Frank Sprenger. The scale bar can be an overlay or applied to the original or a copy of the image. If the image is in color there are multiple color options (the selection is restricted to grayscale choices for grayscale images to retain the original bit depth). The units can be changed from the original embedded scale (i.e. from nm to µm etc.).

The Fancy Summary Label Table macro adds up to 8 lines of statistical summary to an image. There are extensive formatting options including outlines and shadows to help the text stand out against the image. If the image is in color there are multiple color options.

The Fancy Text Labels macro adds up to 8 lines of user created text to an image. There are extensive formatting options includ